# core/specs/api.yaml
openapi: 3.0.0
info:
  title: Infrastructure Toolkit API
  description: REST API for infrastructure monitoring and management
  version: 1.0.0
  contact:
    name: Infrastructure Toolkit
    url: https://github.com/EdFabre/infrastructure-toolkit
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://infra.haymoed.com
    description: Production server
...
paths:
...
  /api/cloudflare/hostnames:
    get:
      tags: [cloudflare]
      summary: List tunnel hostnames
      description: Returns all configured Cloudflare tunnel hostnames
      parameters:
        - name: domain
          in: query
          description: Domain to query
          schema:
            type: string
            enum: [haymoed, ramcyber]
            default: haymoed
      responses:
        '200':
          description: Hostname list
          content:
            application/json:
              schema:
                type: object
                properties:
                  hostnames:
                    type: array
                    items:
                      $ref: '#/components/schemas/CloudflareHostname'
  /api/cloudflare/validate:
    get:
      tags: [cloudflare]
      summary: Validate tunnel configuration
      description: Validates Cloudflare tunnel configuration integrity
      parameters:
        - name: domain
          in: query
          schema:
            type: string
            enum: [haymoed, ramcyber]
            default: haymoed
      responses:
        '200':
          description: Validation result
          content:
            application/json:
              schema:
                type: object
                properties:
                  is_valid:
                    type: boolean
                  errors:
                    type: array
                    items:
                      type: string

  # Pterodactyl endpoints
...
components:
  schemas:
...
    CloudflareHostname:
      type: object
      properties:
        hostname:
          type: string
        service:
          type: string
...